+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Features
+
+- Added API endpoint to see time spent on Github commits.
+
 ## [1.2.0] - 2021-03-14
 
 ### Features